kiratech/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
sqlparse==0.5.3
